feat(coreew): profit-triggered fractional rebalance (COREEW_PROFIT_TRIGGER $100)
When ANY single held ETF leg has unrealized P&L >= threshold, override the
drift gate to an exact equal-weight rebalance that cycle; the threshold
doubles as the min-deviation trade floor. Exact pass trades FRACTIONAL
shares (4 decimals, $1 Alpaca min notional) so sub-share deviations act.
verified live 2026-09-17: QQQ 47->46.7244 (SELL 0.2756 @ 714.635, +$2.99).

- ExecuteEWETF: --profit= option (env COREEW_PROFIT_TRIGGER, default 100,
  0 = disabled); Slack '💰 profit-triggered rebalance' line
- TradeExecutorService: rebalanceEqualWeightWeekly(dryRun, minDriftPct,
  minProfitTrigger) returns profit_triggered; fractional qty in exact pass;
  rebalanceTrim/rebuy/topUp take float qty
- live_trades.quantity + positions_cache.qty: integer -> numeric(12,5);
  reconcilers/sync floatval (EquityService, SyncPositions, EquityController);
  placeOrder sends qty as string
- docs: HANDOFF_CoreEW_rename, TRADING_STRATEGIES, How_System_Works,
  HANDOFF_DailySignal_All3CO, AGENTS.md

+++ swingtrader/backend/app/Console/Commands/SyncPositions.php
@@ -24,7 +24,7 @@ class SyncPositions extends Command
             foreach ($positions as $pos) {
                 PositionCache::create([
                     'symbol' => $pos['symbol'],
-                    'qty' => intval($pos['qty']),
+                    'qty' => floatval($pos['qty']),
                     'avg_entry_price' => floatval($pos['avg_entry_price']),
                     'current_price' => floatval($pos['current_price']),
                     'unrealized_pnl' => floatval($pos['unrealized_pnl'] ?? $pos['unrealized_pl'] ?? 0),
